requirements.txt
Django~=4.0.5
djangorestframework~=3.13.1
Pillow~=9.2.0